+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.0.6] - 2025-11-05
+
+### Added
+- 自定义天空背景系统 for OrderCore维度 (dim103)
+- 黑色太阳和苍白月亮渲染效果
+- 自定义天空盒、雾效和环境颜色控制
+- OrderCoreSkyRenderer with celestial body customization
+
+### Changed
+- WorldProviderOrderCore 增加了天空颜色和环境控制方法
+- 优化了维度的视觉效果和氛围
+
 ## [1.0.0] - 2023-09-15
 
 ### Added